requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
nltk==3.8.1
scikit-learn==1.3.0
openai==0.28.0
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.13.0
Pillow==10.0.0
requests==2.31.0
telebot==0.0.5
tqdm==4.66.1
transformers==4.6.0
urllib3==2.0.4
yarl==1.9.2
